Rename generated libraries.
