+++ jni/CMakeLists.txt
@@ -1,17 +1,17 @@
-add_library(basic_pedestrian_tracker_jni SHARED tracker_jni.cpp)
+add_library(jnibasic_pedestrian_tracker SHARED jnitracker.cpp)
 
 set (CMAKE_C_CFLAGS "${CMAKE_C_CFLAGS} -fPIC")
 set (CMAKE_CXX_CFLAGS "${CMAKE_CXX_CFLAGS} -fPIC")
 
-target_link_libraries (basic_pedestrian_tracker_jni basic_pedestrian_tracker)
+target_link_libraries(jnibasic_pedestrian_tracker basic_pedestrian_tracker)
 
-include_directories (basic_pedestrian_tracker_jni PUBLIC ${PROJECT_SOURCE_DIR}/include)
-include_directories (basic_pedestrian_tracker_jni PUBLIC ${JAVA_HOME}/include)
+include_directories(jnibasic_pedestrian_tracker PUBLIC ${PROJECT_SOURCE_DIR}/include)
+include_directories(jnibasic_pedestrian_tracker PUBLIC ${JAVA_HOME}/include)
 
 if (MSVC)
-    include_directories (basic_pedestrian_tracker_jni PUBLIC ${JAVA_HOME}/include/windows)
+    include_directories(jnibasic_pedestrian_tracker PUBLIC ${JAVA_HOME}/include/windows)
 else ()
-    include_directories (basic_pedestrian_tracker_jni PUBLIC ${JAVA_HOME}/include/linux)
+    include_directories(jnibasic_pedestrian_tracker PUBLIC ${JAVA_HOME}/include/linux)
 endif ()
 
-install (TARGETS basic_pedestrian_tracker_jni DESTINATION lib)
+install(TARGETS jnibasic_pedestrian_tracker DESTINATION lib)
