+++ test/CMakeLists.txt
@@ -1,9 +1,9 @@
 enable_testing()
 
-add_executable(pedestrian_tracker_test EXCLUDE_FROM_ALL pedestrian_tracker_test.cpp)
-include_directories(pedestrian_tracker_test PUBLIC ${PROJECT_SOURCE_DIR}/include)
-target_link_libraries(pedestrian_tracker_test basic_pedestrian_tracker opencv_core opencv_highgui opencv_imgproc)
-add_test(NAME pedestrian_tracker_test COMMAND pedestrian_tracker_test ${TEST_VIDEO_PATH})
+add_executable(tracker_test EXCLUDE_FROM_ALL tracker_test.cpp)
+include_directories(tracker_test PUBLIC ${PROJECT_SOURCE_DIR}/include)
+target_link_libraries(tracker_test basic_pedestrian_tracker opencv_core opencv_highgui opencv_imgproc)
+add_test(NAME tracker_test COMMAND tracker_test ${TEST_VIDEO_PATH})
 
 add_executable(lib_test EXCLUDE_FROM_ALL lib_test.cpp)
 include_directories(lib_test PUBLIC ${PROJECT_SOURCE_DIR}/include)
